Refactor GameTimer and GameToolBar components to remove pause functionality; simplify GameTimer layout and update GameToolBar documentation.
