+++ SudokuApp/components/GameTimer.js
@@ -1,19 +1,14 @@
 import React from 'react';
-import { View, StyleSheet, Text, TouchableOpacity, Platform } from 'react-native';
-import { useGameContext, ACTIONS } from '../contexts/GameContext';
+import { View, StyleSheet, Text, Platform } from 'react-native';
+import { useGameContext } from '../contexts/GameContext';
 import LabeledBadge from './LabeledBadge';
-import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons';
 
 /**
- * GameTimer component displaying elapsed time and pause button
+ * GameTimer component displaying elapsed time
  * Positioned on the right side of the GameTopStrip
  */
 const GameTimer = () => {
-  const { elapsedSeconds, formatTime, dispatch, theme, isPaused } = useGameContext();
-
-  const handlePausePress = () => {
-    dispatch({ type: ACTIONS.PAUSE_GAME });
-  };
+  const { elapsedSeconds, formatTime, theme } = useGameContext();
 
   return (
     <LabeledBadge
@@ -28,21 +23,6 @@ const GameTimer = () => {
         ]}>
           {formatTime(elapsedSeconds)}
         </Text>
-        
-        {/* Pause button integrated in badge */}
-        <TouchableOpacity 
-          style={styles.pauseButton}
-          onPress={handlePausePress}
-          disabled={isPaused}
-          accessibilityLabel="Pause Game"
-          hitSlop={{ top: 8, bottom: 8, left: 8, right: 8 }}
-        >
-          <MaterialCommunityIcons
-            name="pause-circle"
-            size={21}
-            color={theme.colors.numberPad.text}
-          />
-        </TouchableOpacity>
       </View>
     </LabeledBadge>
   );
@@ -53,7 +33,6 @@ const styles = StyleSheet.create({
     alignItems: 'center',
   },
   timerInnerContainer: {
-    flexDirection: 'row',
     alignItems: 'center',
     justifyContent: 'center',
     height: 24,
@@ -63,15 +42,8 @@ const styles = StyleSheet.create({
     fontWeight: 'bold',
     textAlign: 'center',
     letterSpacing: 1,
-    marginRight: 4, // Space between time and pause button
     fontFamily: Platform.select({ ios: 'Menlo', android: 'monospace', default: 'monospace' }),
     width: 60, // Fixed width for timer
-  },
-  pauseButton: {
-    width: 24,
-    height: 24,
-    justifyContent: 'center',
-    alignItems: 'center',
   },
 });
